Scale ggseg plot axes.

adapt_scales returns a list of coordinate breaks and labels for axes or axes label manipulation of the ggseg brain atlases.

adapt_scales(geobrain, position = "dispersed", aesthetics = "labs")

Arguments

geobrain
a data.frame containing atlas information.

position
String choosing how to view the data. Either "dispersed"[default] or "stacked".

aesthetics
String of which aesthetics to adapt scale of, either "x","y", or "labs".

Value

nested list