Desikan-Killiany Cortical Atlas - tri-surf mesh Mesh data for the Desikan-Killany Cortical atlas, with 40 regions in on the cortical surface of the brain.

A nested tibble for all available surfaces and hemispheres

data(dkt_3d)

Format

A tibble with 4 observations and a nested data.frame

surf

type of surface (`inflated` or `white`)

hemi

hemisphere (`left`` or `right`)

data

data.frame of necessary variables for plotting

atlas

String. atlas name

roi

numbered region from surface

annot

concatenated region name

label

label `hemi_annot` of the region

mesh

list of meshes in two lists: vb and it

acronym

abbreviated name of annot

lobe

lobe localization

area

name of area in full

colour

HEX colour of region

References

Fischl et al. (2004) Cerebral Cortex 14:11-22 (PubMed)

See also

Examples

data(dkt_3d)