ggseg: Visualise brain atlas data with ggplot2

Provides a ggplot2 geom and position for visualising brain region data on cortical, subcortical, and white matter tract atlases. Brain atlas geometries are stored as simple features (sf), enabling seamless integration with the ggplot2 ecosystem including faceting, custom scales, and themes.

`ggseg()` is defunct as of version 2.0.0. Use `ggplot() + geom_brain()` instead.

Usage

ggseg(...)

Arguments

...

Ignored.

Value

Does not return; always raises an error.

Details

The main entry point is [geom_brain()], which accepts a `ggseg_atlas` object and optional user data. Use [position_brain()] to control the layout of brain slices/views.

See also

Author

Maintainer: Athanasia Mo Mowinckel a.m.mowinckel@psykologi.uio.no (ORCID)

Authors:

Didac Vidal-Piñeiro d.v.pineiro@psykologi.uio.no (ORCID)

Other contributors:

Ramiro Magno ramiro.magno@gmail.com (ORCID) [contributor]

Center for Lifespan Changes in Brain and Cognition, University of Oslo, Norway [copyright holder]

Examples

if (FALSE) { # \dontrun{
ggseg()
} # }